Add analytics package
