+++ projects/js-packages/analytics/CHANGELOG.md
@@ -1,0 +1,3 @@
+# Changelog
+
+### This is a list detailing changes for the Jetpack RNA Components package releases.
