+++ projects/js-packages/analytics/changelog/initial-version
@@ -1,0 +1,4 @@
+Significance: minor
+Type: added
+
+Initial release of jetpack-api package
