+++ projects/js-packages/analytics/test-main.jsx
@@ -1,0 +1,7 @@
+/**
+ * External dependencies
+ */
+import Enzyme from 'enzyme';
+import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+
+Enzyme.configure( { adapter: new Adapter() } );
