+++ projects/plugins/jetpack/package.json
@@ -62,6 +62,7 @@
 		"@automattic/calypso-color-schemes": "2.1.1",
 		"@automattic/components": "1.0.0-alpha.3",
 		"@automattic/format-currency": "1.0.0-alpha.0",
+		"@automattic/jetpack-analytics": "workspace:^0.1.0-alpha",
 		"@automattic/jetpack-api": "workspace:^0.1.0",
 		"@automattic/jetpack-components": "workspace:^0.2.2-alpha",
 		"@automattic/jetpack-connection": "workspace:^0.6.0-alpha",
